+++ packages/blocks/core/CHANGELOG.md
@@ -5,7 +5,7 @@
 - Added support for the `Navigation Link Block` by Blockera. [[🔗 Link](https://community.blockera.ai/feature-request-1rsjg2ck/post/supporting-blocks-inside-navigation-block-MIcY979kIVCxkvU)]
 - Added support for the `Home Link Block` by Blockera. [[🔗 Link](https://community.blockera.ai/feature-request-1rsjg2ck/post/supporting-blocks-inside-navigation-block-MIcY979kIVCxkvU)]
 - Added support for the `Submenu Block` and its inner blocks by Blockera. [[🔗 Link](https://community.blockera.ai/feature-request-1rsjg2ck/post/supporting-blocks-inside-navigation-block-MIcY979kIVCxkvU)]
-
+- Added support for the `Tag Cloud Block` by Blockera.
 
 ### Bug Fixes
 
@@ -16,6 +16,7 @@
 - Added automated test for `Navigation Link Block` to check that the `Back to Parent` navigation buttons work correctly.
 - Added automated test for `Home Link Block` to check that the `Back to Parent` navigation buttons work correctly.
 - Added automated test for `Submenu Block` to check that the `Back to Parent` navigation buttons work correctly.
+- Added automated test to verify functionality of the `Tag Cloud Block` and its inner blocks.
 
 ### Miscellaneous